feat(frontend): RH dashboard Maquettes 4/5/6 — affectations + XAI + historique
Squashed flat onto main (post PR #13 merge). Contains:

Frontend (PR #14 scope):
- New pages: /dashboard/rh/affectations and /dashboard/rh/historique
- Components: AffectationCard (XAI accordion), AffectationTable, ScoreBreakdown,
  WeightSliders, GenerationForm (session/programme/sliders state machine)
- Hook: useGenerationPoller (SWR-based task polling)
- API layer: lib/api/affectations.ts (sessions, programmes, affectations)
- Types: lib/types/api.ts extended for affectations/sessions/programmes
- Vitest unit tests for all new components
- Nav: enable affectations/historique entries in rhNav

Backend fixes (post initial QA):
- fix(algo): eager-load Professeur.user to avoid MissingGreenlet in Celery worker
- fix(algo): compute annees_exp from annee_debut/annee_fin
- fix(api): allow rh role to PUT ponderations (Maquette 4)
- fix(db): seed_affectation_demo — idempotent, real etape IDs, no emoji on Windows
- fix(algo): lazy-import sentence-transformers to avoid blocking uvicorn startup

Frontend fixes:
- fix(frontend): coerce Decimal API strings to numbers (fix toFixed crash on weights and scores)

Docs:
- Plan and spec for RH affectations frontend (Maquettes 4/5/6)

+++ backend/app/schemas/affectation.py
@@ -52,10 +52,10 @@ class SessionCreate(BaseModel):
 
 class PonderationsOut(BaseModel):
     session_id: int
-    w1: Decimal
-    w2: Decimal
-    w3: Decimal
-    w4: Decimal
+    w1: float
+    w2: float
+    w3: float
+    w4: float
     xai_actif: bool
     mis_a_jour_le: datetime
 
@@ -84,11 +84,11 @@ class AffectationOut(BaseModel):
     session_id: int
     professeur_id: int
     cours_id: int
-    score_total: Decimal
-    score_comp: Decimal
-    score_exp: Decimal
-    score_hist: Decimal
-    score_sem: Decimal
+    score_total: float
+    score_comp: float
+    score_exp: float
+    score_hist: float
+    score_sem: float
     justification: Optional[str]
     statut: AffectationStatut
     valide_par_user_id: Optional[int]
